Back/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
defusedxml==0.7.1
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.27.0
idna==3.6
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.6
psycopg2==2.9.9
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.17.2
PySocks==1.7.1
python-dotenv==1.0.0
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
setuptools==69.0.3
shellingham==1.5.4
sniffio==1.3.0
SQLAlchemy==2.0.25
sqlmodel==0.0.19
starlette==0.37.2
typer==0.12.3
typing_extensions==4.9.0
ujson==5.10.0
urllib3==2.1.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
